chore: added tests
