+++ .gitignore
@@ -3,3 +3,5 @@ scratches/
 __pycache__/
 dist/
 .DS_Store
+.mypy_cache/
+.ruff_cache/
