--- altb/common.py
@@ -1,5 +1,0 @@
-from rich.console import Console
-
-
-console = Console()
-error_console = Console(stderr=True, style="bold red")
