+++ altb/common/console.py
@@ -1,0 +1,4 @@
+from rich.console import Console
+
+console = Console()
+error_console = Console(stderr=True, style="bold red")
