+++ pyproject.toml
@@ -24,9 +24,18 @@ natsort = "^8.3.1"
 [tool.poetry.dev-dependencies]
 ipdb = "^0.13.9"
 
+[tool.poetry.group.dev.dependencies]
+pytest = "^7.3.0"
+ruff = "^0.0.261"
+mypy = "^1.2.0"
+types-pyyaml = "^6.0.12.9"
+
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
 altb = 'altb.main:main'
+
+[tool.ruff]
+line-length = 120
